results.versions/v0.requirements.txt
absl-py==2.1.0
accelerate==0.33.0
aiohappyeyeballs==2.3.5
aiohttp==3.10.2
aiosignal==1.3.1
annotated-types==0.7.0
anytree==2.12.1
astunparse==1.6.3
async-timeout==4.0.3
attrs==24.2.0
beautifulsoup4==4.12.3
bert-score==0.3.13
bioc==2.1
blis==0.7.11
boto3==1.34.157
botocore==1.34.157
cachetools==5.4.0
catalogue==2.0.10
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
cloudpathlib==0.18.1
cloudpickle==3.0.0
confection==0.1.5
conllu==4.5.3
contourpy==1.1.1
cycler==0.12.1
cymem==2.0.8
datasets==2.20.0
decorator==5.1.1
deprecated==1.2.14
dill==0.3.8
dm-tree==0.1.8
docopt==0.6.2
editdistance==0.8.1
filelock==3.15.4
flair==0.14.0
flatbuffers==24.3.25
fonttools==4.53.1
frozenlist==1.4.1
fsspec==2024.5.0
ftfy==6.2.3
gast==0.4.0
gdown==5.2.0
google-auth==2.33.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.65.4
h5py==3.11.0
huggingface-hub==0.24.5
idna==3.7
importlib-metadata==8.2.0
importlib-resources==6.4.0
intervaltree==3.1.0
jieba==0.42.1
jinja2==3.1.4
jmespath==1.0.1
joblib==1.4.2
jsonlines==4.0.0
keras==2.10.0
keras-preprocessing==1.1.2
kiwisolver==1.4.5
langcodes==3.4.0
langdetect==1.0.9
language-data==1.2.0
language-tool-python==2.8
lemminflect==0.2.3
libclang==18.1.1
lru-dict==1.3.0
lxml==5.2.2
marisa-trie==1.2.0
markdown==3.6
markdown-it-py==3.0.0
markupsafe==2.1.5
matplotlib==3.7.5
mdurl==0.1.2
more-itertools==10.4.0
mpld3==0.5.10
mpmath==1.3.0
multidict==6.0.5
multiprocess==0.70.16
murmurhash==1.0.10
networkx==3.1
nltk==3.8.1
num2words==0.5.13
numpy==1.24.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.6.20
nvidia-nvtx-cu12==12.1.105
oauthlib==3.2.2
openhownet==2.0
opt-einsum==3.3.0
packaging==24.1
pandas==2.0.3
patsy==0.5.6
pillow==10.4.0
pinyin==0.4.0
pptree==3.1
preshed==3.0.9
protobuf==3.19.6
psutil==6.0.0
pyarrow==17.0.0
pyarrow-hotfix==0.6
pyasn1==0.6.0
pyasn1-modules==0.4.0
pycld2==0.41
pydantic==2.8.2
pydantic-core==2.20.1
pygments==2.18.0
pyparsing==3.1.2
pysocks==1.7.1
python-dateutil==2.9.0.post0
pytorch-revgrad==0.2.0
pytz==2024.1
pyyaml==6.0.2
regex==2024.7.24
requests==2.32.3
requests-oauthlib==2.0.0
rich==13.7.1
rsa==4.9
s3transfer==0.10.2
safetensors==0.4.4
scikit-learn==1.3.2
scipy==1.10.1
seaborn==0.13.2
segtok==1.5.11
semver==3.0.2
sentencepiece==0.2.0
shellingham==1.5.4
six==1.16.0
smart-open==7.0.4
sortedcontainers==2.4.0
soupsieve==2.5
spacy==3.7.5
spacy-legacy==3.0.12
spacy-loggers==1.0.5
sqlitedict==2.1.0
srsly==2.4.8
statsmodels==0.14.1
sympy==1.13.1
tabulate==0.9.0
tensorboard==2.10.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.10.0
tensorflow-estimator==2.10.0
tensorflow-hub==0.16.1
tensorflow-io-gcs-filesystem==0.34.0
tensorflow-privacy==0.7.0
tensorflow-probability==0.18.0
termcolor==2.4.0
terminaltables==3.1.10
textattack==0.3.10
tf-keras==2.15.0
thinc==8.2.5
threadpoolctl==3.5.0
tokenizers==0.19.1
torch==2.4.0
tqdm==4.66.5
transformer-smaller-training-vocab==0.4.0
transformers==4.44.0
triton==3.0.0
typer==0.12.3
typing-extensions==4.12.2
tzdata==2024.1
urllib3==1.26.19
wasabi==1.1.3
wcwidth==0.2.13
weasel==0.4.1
werkzeug==3.0.3
wikipedia-api==0.6.0
word2number==1.1
wrapt==1.16.0
xxhash==3.4.1
yarl==1.9.4
zipp==3.19.2